refactor(release): unify updater manifest merge scripts
