--- scripts/merge-mac-update-manifests.ts
@@ -1,51 +1,0 @@
-import { readFileSync, writeFileSync } from "node:fs";
-import { resolve } from "node:path";
-import { fileURLToPath } from "node:url";
-
-import {
-  mergeUpdateManifests,
-  parseUpdateManifest,
-  serializeUpdateManifest,
-  type UpdateManifest,
-} from "./lib/update-manifest.ts";
-
-type MacUpdateManifest = UpdateManifest;
-
-export function parseMacUpdateManifest(raw: string, sourcePath: string): MacUpdateManifest {
-  return parseUpdateManifest(raw, sourcePath, "macOS");
-}
-
-export function mergeMacUpdateManifests(
-  primary: MacUpdateManifest,
-  secondary: MacUpdateManifest,
-): MacUpdateManifest {
-  return mergeUpdateManifests(primary, secondary, "macOS");
-}
-
-export function serializeMacUpdateManifest(manifest: MacUpdateManifest): string {
-  return serializeUpdateManifest(manifest, {
-    platformLabel: "macOS",
-  });
-}
-
-function main(args: ReadonlyArray<string>): void {
-  const [arm64PathArg, x64PathArg, outputPathArg] = args;
-  if (!arm64PathArg || !x64PathArg) {
-    throw new Error(
-      "Usage: node scripts/merge-mac-update-manifests.ts <latest-mac.yml> <latest-mac-x64.yml> [output-path]",
-    );
-  }
-
-  const arm64Path = resolve(arm64PathArg);
-  const x64Path = resolve(x64PathArg);
-  const outputPath = resolve(outputPathArg ?? arm64PathArg);
-
-  const arm64Manifest = parseMacUpdateManifest(readFileSync(arm64Path, "utf8"), arm64Path);
-  const x64Manifest = parseMacUpdateManifest(readFileSync(x64Path, "utf8"), x64Path);
-  const merged = mergeMacUpdateManifests(arm64Manifest, x64Manifest);
-  writeFileSync(outputPath, serializeMacUpdateManifest(merged));
-}
-
-if (process.argv[1] && resolve(process.argv[1]) === fileURLToPath(import.meta.url)) {
-  main(process.argv.slice(2));
-}
